[sprint4][export file type use public for type of store include reducer,action,component,saga]
