+++ feproduct/src/store/redux/newsUser.tsx
@@ -1,45 +1,18 @@
 import React from 'react';
 import {updateObject} from "../share/utility";
 import * as actionTypes from "../action/actiontypes";
+import { tsInitialState,tsActionNews } from "../../types/newsType"
 
-export interface tsInitialState {
-    listNews:  Array<tscategoryNews>,
-}
+
 
 const initialState: tsInitialState = {
     listNews: [],
 }
-export type  images = {
-    _id: string,
-    public_id: string,
-    url:string
-}
-export type News = {
-    _id: string,
-    title: string,
-    is_show: Boolean,
-    content: string,
-    category_News: string,
-    createdAt: string,
-    images: Array<images>
-}
-export type tscategoryNews = {
-    _id: string,
-    name: string,
-    createdAt: string,
-    updatedAt: string,
-    news: Array<News>
-    type: string
-}
-type tsAction = {
-    data: Array<tscategoryNews>,
-    type: string
-}
-const dataNews = (state:tsInitialState,action:tsAction) => {
+const dataNews = (state:tsInitialState, action:tsActionNews) => {
     // console.log("actionnews",action.data);
     return updateObject(state,{listNews: action.data})
 }
-const newsReducer = (state = initialState,action:tsAction) => {
+const newsReducer = (state = initialState,action:tsActionNews) => {
     switch (action.type)
     {
         case actionTypes.DATA_NEWS_USER:
