+++ feproduct/src/types/productType.tsx
@@ -1,0 +1,46 @@
+export type  Images = {
+    _id: string,
+    public_id: string,
+    url:string
+}
+export type Options = {
+    _id:string,
+    type: string,
+    size: string,
+    code: string,
+    price: number,
+    quantity: number,
+    specifications: string,
+}
+export type Product = {
+    _id: string,
+    Product_name: string,
+    images: Array<Images>,
+    description: string,
+    options: Array<Options>,
+    totalAmount: number,
+    quantityCart: number,
+}
+export interface actionTypeCart {
+    type: string,
+    data: Product,
+    id: number,
+    calculation:string
+}
+export interface tsInitialState {
+    status: string,
+    cart: Product[],
+    quantityCart: number,
+    totalMoney: number
+}
+export type tsCategoryProduct = {
+    updatedAt: string,
+    createdAt: string,
+    name: string,
+    _id: string,
+    product: Array<Product>
+} 
+export interface fetchComments {
+    data: Product[],
+    totalPage: number
+}
